+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Correct victron energy manager data persistence
+
 ## 2.13.1 (2023-09-24)
 
 * (theimo1221) Add Battery Data to energyCalculation